# perf: optimize Flac decoding