… lib/src/wav/wav_encoder.dart