It's 2s faster on my machine when I compile a basic app.  From `3.4s` to `1.3s`.

The main gains :
- Send to the MD5 input the full list of samples instead of sending samples one by one
- Optimize the WAV encoder. Use a Uint32List. Gain of almost 1s
- Merge the `BitReader` into `Buffer` and improve the `Buffer` logic.
- Add the function `readBit()` that is more performant than `readingBits(1)`


And some refactoring to make the code clearer and some unit tests